% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run TAF Analysis}
\usage{
run(force = FALSE)
}
\arguments{
\item{force}{whether to remove existing directory.}
}
\description{
Run entire TAF analysis, both bootstrap procedure and scripts.
}
\note{
This function creates a new working directory with the same name as the
package, before running the bootstrap procedure (\code{taf.bootstrap}) and
TAF scripts (\code{sourceAll}).
}
\examples{
run(force=TRUE)

}
\seealso{
\code{\link[icesTAF]{taf.bootstrap}} sets up data files and software.

\code{\link[icesTAF]{sourceAll}} runs all TAF scripts.

\code{\link[icesTAF]{icesTAF-package}} gives an overview of the icesTAF
package.
}
